\name{remotecall}
\alias{remotecall}
\title{Remote call server (Demonstration)}
\description{
  General information about the remotecall client/server interface.
}
\details{
    The package remotecall is a simple demonstration package
    for a remote call interface via TCP/IP.

    See \code{library(help=remotecall)} for a list of all available functions.

}
\author{Thorsten Rangwich}
